Move LayoutTab to its own .cpp and .h files
