+++ foo_ui_columns/foo_ui_columns.vcxproj
@@ -447,6 +447,7 @@
     <ClCompile Include="notification_area.cpp" />
     <ClCompile Include="rebar_band.cpp" />
     <ClCompile Include="splitter_utils.cpp" />
+    <ClCompile Include="tab_layout.cpp" />
     <ClCompile Include="tab_layout_misc.cpp" />
     <ClCompile Include="tab_pview_artwork.cpp" />
   </ItemGroup>
@@ -511,6 +512,7 @@
     <ClInclude Include="notification_area.h" />
     <ClInclude Include="rebar_band.h" />
     <ClInclude Include="splitter_utils.h" />
+    <ClInclude Include="tab_layout.h" />
     <ClInclude Include="tab_layout_misc.h" />
   </ItemGroup>
   <ItemGroup>
